mv mut_output and push_changes fixt to integration conf
